Move transform package out from ingestor package
It's used by collector and ingestor.
